[#163] 작업정의 상세 페이지에 작업 제출 버튼 구현
