+++ web/src/components/TaskDefinitionDetailModal.tsx
@@ -12,12 +12,10 @@ import {
   InputLabel,
   Select,
   MenuItem,
-  Snackbar,
-  Alert,
 } from '@mui/material';
 import { PlayArrow as PlayArrowIcon, Add as AddIcon } from '@mui/icons-material';
-import { TaskDefinition, ErrorResponse, submitJob } from '../api';
-import { useNavigate } from 'react-router-dom';
+import { TaskDefinition } from '../api';
+import JobSubmitModal from './JobSubmitModal';
 
 interface TaskDefinitionDetailModalProps {
   open: boolean;
@@ -32,24 +30,18 @@ const TaskDefinitionDetailModal: React.FC<TaskDefinitionDetailModalProps> = ({
   taskDefinition,
   onCreateVersion,
 }) => {
-  const navigate = useNavigate();
-  const [isSubmitModalOpen, setIsSubmitModalOpen] = useState(false);
-  const [jobName, setJobName] = useState('');
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [successMessage, setSuccessMessage] = useState<string | null>(null);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
+  const [isJobSubmitModalOpen, setIsJobSubmitModalOpen] = useState(false);
 
   if (!taskDefinition) return null;
 
   const env = taskDefinition.env ? JSON.parse(taskDefinition.env) : [];
 
   const handleSubmitClick = () => {
-    setIsSubmitModalOpen(true);
+    setIsJobSubmitModalOpen(true);
   };
 
-  const handleSubmitClose = () => {
-    setIsSubmitModalOpen(false);
-    setJobName('');
+  const handleJobSubmitModalClose = () => {
+    setIsJobSubmitModalOpen(false);
   };
 
   const handleCreateVersion = () => {
@@ -57,48 +49,6 @@ const TaskDefinitionDetailModal: React.FC<TaskDefinitionDetailModalProps> = ({
       onCreateVersion(taskDefinition);
       onClose();
     }
-  };
-
-  const handleSubmitConfirm = async () => {
-    if (!jobName.trim()) {
-      setErrorMessage('작업 이름을 입력해주세요.');
-      return;
-    }
-
-    try {
-      setIsSubmitting(true);
-      setErrorMessage(null);
-      
-      const result = await submitJob({
-        task_definition_id: taskDefinition.id,
-        job_name: jobName.trim(),
-      });
-
-      if (result.response instanceof ErrorResponse) {
-        setErrorMessage(`작업 제출에 실패했습니다: ${result.response.message || '알 수 없는 오류'}`);
-        console.error('Failed to submit job:', result.response.error_code, result.response.message);
-      } else {
-        setSuccessMessage(`작업 "${jobName}"이 성공적으로 제출되었습니다.`);
-        handleSubmitClose();
-        // 2초 후 작업 목록 페이지로 이동
-        setTimeout(() => {
-          navigate('/jobs');
-        }, 2000);
-      }
-    } catch (error) {
-      setErrorMessage('작업 제출 중 오류가 발생했습니다.');
-      console.error('Failed to submit job:', error);
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const handleCloseSuccessAlert = () => {
-    setSuccessMessage(null);
-  };
-
-  const handleCloseErrorAlert = () => {
-    setErrorMessage(null);
   };
 
   return (
@@ -215,57 +165,12 @@ const TaskDefinitionDetailModal: React.FC<TaskDefinitionDetailModalProps> = ({
         </DialogActions>
       </Dialog>
 
-      <Dialog
-        open={isSubmitModalOpen}
-        onClose={handleSubmitClose}
-      >
-        <DialogTitle>작업 제출</DialogTitle>
-        <DialogContent>
-          <TextField
-            autoFocus
-            margin="dense"
-            label="작업 이름"
-            fullWidth
-            value={jobName}
-            onChange={(e) => setJobName(e.target.value)}
-            required
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleSubmitClose} disabled={isSubmitting}>취소</Button>
-          <Button
-            onClick={handleSubmitConfirm}
-            variant="contained"
-            disabled={!jobName.trim() || isSubmitting}
-          >
-            {isSubmitting ? '제출 중...' : '제출'}
-          </Button>
-        </DialogActions>
-      </Dialog>
-
-      {/* 성공 알림 */}
-      <Snackbar
-        open={!!successMessage}
-        autoHideDuration={6000}
-        onClose={handleCloseSuccessAlert}
-        anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-      >
-        <Alert onClose={handleCloseSuccessAlert} severity="success" sx={{ width: '100%' }}>
-          {successMessage}
-        </Alert>
-      </Snackbar>
-
-      {/* 에러 알림 */}
-      <Snackbar
-        open={!!errorMessage}
-        autoHideDuration={6000}
-        onClose={handleCloseErrorAlert}
-        anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-      >
-        <Alert onClose={handleCloseErrorAlert} severity="error" sx={{ width: '100%' }}>
-          {errorMessage}
-        </Alert>
-      </Snackbar>
+      {/* 작업 제출 모달 */}
+      <JobSubmitModal
+        open={isJobSubmitModalOpen}
+        onClose={handleJobSubmitModalClose}
+        taskDefinition={taskDefinition}
+      />
     </>
   );
 };
